Test the seqspec index command
We have some read_ids that are just accession and some that are
accession.fastq.gz and this causes confusion for generating the tool
command lines.

I thought I should save some example seqspecs and test that index
generates the correct commands in with either type of read id.

Also to get this to work I decided to use importlib.resources which
required making my test/data directory look like a python module.